+++ packages/api/src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './getNextParams';
 export * from './checkObjEqual';
-export * from './resolveDefaultParams';
