+++ packages/evmApi/src/resolvers/native/getLogsByAddress.ts
@@ -1,7 +1,8 @@
-import { ApiPaginatedOptions, ApiPaginatedResolver, resolveDefaultChain } from '@moralisweb3/api';
+import { ApiPaginatedOptions, ApiPaginatedResolver } from '@moralisweb3/api';
 import { EvmChainish, EvmAddressish, EvmAddress, EvmTransactionLog, Camelize, toCamelCase } from '@moralisweb3/core';
 import { BASE_URL } from '../../EvmApi';
 import { operations } from '../../generated/types';
+import { EvmChainResolver } from '../EvmChainResolver';
 
 type operation = 'getLogsByAddress';
 
@@ -31,7 +32,7 @@ export const getLogsByAddressResolver = new ApiPaginatedResolver({
   resultToJson: (data) => data?.map((log) => log.toJSON()),
   parseParams: (params: Params): ApiParams => ({
     ...params,
-    chain: resolveDefaultChain(params.chain).apiHex,
+    chain: EvmChainResolver.resolve(params.chain).apiHex,
     address: EvmAddress.create(params.address).lowercase,
   }),
 });
