+++ packages/evmApi/src/resolvers/token/getAllTokenIds.ts
@@ -1,7 +1,8 @@
-import { ApiPaginatedResolver, ApiPaginatedOptions, resolveDefaultChain } from '@moralisweb3/api';
+import { ApiPaginatedResolver, ApiPaginatedOptions } from '@moralisweb3/api';
 import { EvmChainish, EvmAddressish, EvmAddress, EvmNFT, Camelize } from '@moralisweb3/core';
 import { BASE_URL } from '../../EvmApi';
 import { operations } from '../../generated/types';
+import { EvmChainResolver } from '../EvmChainResolver';
 
 type operation = 'getAllTokenIds';
 
@@ -41,7 +42,7 @@ export const getAllTokenIdsResolver = new ApiPaginatedResolver({
   apiToResult: (data: ApiResult, params: Params) =>
     data.result?.map((nft) => ({
       token: new EvmNFT({
-        chain: resolveDefaultChain(params.chain),
+        chain: EvmChainResolver.resolve(params.chain),
         contractType: nft.contract_type,
         tokenAddress: nft.token_address,
         tokenId: nft.token_id,
@@ -66,7 +67,7 @@ export const getAllTokenIdsResolver = new ApiPaginatedResolver({
     })),
   parseParams: (params: Params): ApiParams => ({
     ...params,
-    chain: resolveDefaultChain(params.chain).apiHex,
+    chain: EvmChainResolver.resolve(params.chain).apiHex,
     address: EvmAddress.create(params.address).lowercase,
   }),
 });
