+++ packages/evmApi/src/resolvers/token/getTokenMetadataBySymbol.ts
@@ -1,7 +1,8 @@
-import { ApiResolver, resolveDefaultChain } from '@moralisweb3/api';
+import { ApiResolver } from '@moralisweb3/api';
 import { Erc20Token, EvmChainish, Camelize, toCamelCase } from '@moralisweb3/core';
 import { BASE_URL } from '../../EvmApi';
 import { operations } from '../../generated/types';
+import { EvmChainResolver } from '../EvmChainResolver';
 
 type operation = 'getTokenMetadataBySymbol';
 
@@ -22,7 +23,7 @@ export const getTokenMetadataBySymbolResolver = new ApiResolver({
       const erc20token = new Erc20Token({
         ...toCamelCase(token),
         contractAddress: token.address,
-        chain: resolveDefaultChain(params.chain),
+        chain: EvmChainResolver.resolve(params.chain),
       });
       return {
         token: erc20token,
@@ -33,6 +34,6 @@ export const getTokenMetadataBySymbolResolver = new ApiResolver({
   resultToJson: (data) => data.map((result) => ({ ...result, token: result.token.toJSON() })),
   parseParams: (params: Params): ApiParams => ({
     ...params,
-    chain: resolveDefaultChain(params.chain).apiHex,
+    chain: EvmChainResolver.resolve(params.chain).apiHex,
   }),
 });
